move settings.py and stations_data_export.py
